Move selectionType to dedicated file
